+++ test/test_ecc_base.jl
@@ -245,6 +245,7 @@ const code_instance_args = Dict(
     :QuantumReedMuller => [3, 4, 5],
     :Triangular488 => [3, 5, 7, 9, 11],
     :Triangular666 => [3, 5, 7, 9, 11]
+    :DelfosseReichardtRepCode => [4, 6, 8, 10]
 )
 
 function all_testablable_code_instances(;maxn=nothing)